refactor: move text editing controller from FormulaeStore (#1002)
### Related to
NA

### Context
Move FormulaeTextEditingController to FormulaeController in an effort of making the FormulaeStore solely responsible for holding the StateMachine and the result of the parsed formula. In a follow-up PR the actual state machine definition will be moved into the FormulaeStore, we might even rename the class to FormulaeStateMachine, we'll see.